// src/api/server-schedule-data.model.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
import { ServerFoundationInfoModel } from "../state/schedule-data/foundation-info/foundation-info.model";
import { MonthInfoModel } from "../state/schedule-data/month-info/month-info.model";
import { ShiftTypesDict } from "../state/schedule-data/shifts-types/shift-types.model";
import { WorkersInfoModel } from "../state/schedule-data/worker-info/worker-info.model";
import { WorkerShiftsModel } from "../state/schedule-data/workers-shifts/worker-shifts.model";

export interface ServerScheduleDataModel {
  schedule_info: MonthInfoModel;
  month_info: ServerFoundationInfoModel;
  employee_info: WorkersInfoModel;
  shifts: WorkerShiftsModel;
  shift_types: ShiftTypesDict;
  isAutoGenerated: boolean;
  isCorrupted: boolean;
}